requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
alembic
annotated-types
anthropic
anyio
APScheduler
argon2-cffi
argon2-cffi-bindings
arrow
astroid
asttokens
async-lru
async-timeout
asyncer
attrs
babel
backcall
backoff
backports.tarfile
beautifulsoup4
bleach
boto3
botocore
build
CacheControl
cachetools
certifi
cffi
charset-normalizer
cleo
click
cloudpickle
cmake
colorama
colorlog
comm
contourpy
courlan
crashtest
cryptography
cycler
dataclasses-json
datasets
dateparser
debugpy
decorator
defusedxml
Deprecated
dill
diskcache
distlib
distro
dnspython
dspy
duckduckgo_search
dulwich
email_validator
executing
fastapi
fastapi-cli
fastapi-sso
fastjsonschema
filelock
filetype
fireworks-ai
fonttools
fqdn
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
gunicorn
h11
html2text
htmldate
httpcore
httplib2
httptools
httpx
httpx-sse
httpx-ws
huggingface-hub
idna
importlib_metadata
iniconfig
installer
ipykernel
ipython
ipython-genutils
ipywidgets
isoduration
isort
jaraco.classes
jaraco.context
jaraco.functools
jedi
Jinja2
jiter
jmespath
joblib
json5
json_repair
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter-server-mathjax
jupyter-server-ydoc
jupyter-ydoc
jupyter_client
jupyter_core
jupyter_server
jupyter_server_fileid
jupyter_server_terminals
jupyterlab
jupyterlab_git
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
jusText
keyring
kiwisolver
langchain
langchain-anthropic
langchain-aws
langchain-community
langchain-core
langchain-fireworks
langchain-google-genai
langchain-ollama
langchain-openai
langchain-text-splitters
langsmith
lit
litellm
lmnr
lxml
lxml_html_clean
magicattr
MainContentExtractor
Mako
markdown-it-py
markdownify
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mccabe
mdurl
mistune
monotonic
more-itertools
mpmath
msgpack
multidict
multiprocess
mypy-extensions
nbclassic
nbclient
nbconvert
nbdime
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-curand-cu11
nvidia-nvtx-cu11
oauthlib
ollama
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-langchain
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-threading
opentelemetry-instrumentation-urllib3
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-semantic-conventions-ai
opentelemetry-util-http
optuna
orjson
outcome
overrides
packaging
pandas
pandocfilters
parso
pexpect
pickleshare
pillow
pkginfo
platformdirs
playwright
plotly
pluggy
poetry
poetry-core
poetry-plugin-export
posthog
primp
prometheus_client
prompt_toolkit
propcache
proto-plus
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pycrdt
pycrdt-websocket
pydantic
pydantic-settings
pydantic_core
pyee
PyGithub
Pygments
PyJWT
pylint
PyNaCl
pyparsing
pyproject_hooks
pyrsistent
PySocks
pytest
python-dateutil
python-dotenv
python-json-logger
python-multipart
pytz
PyYAML
pyzmq
RapidFuzz
redis
referencing
regex
requests
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rich-toolkit
rpds-py
rq
rsa
s3transfer
scikit-learn
scipy
seaborn
selenium
Send2Trash
shellingham
six
smmap
smolagents
sniffio
sortedcontainers
soupsieve
SQLAlchemy
sqlite-anyio
stack-data
starlette
sympy
tenacity
terminado
threadpoolctl
tiktoken
tinycss2
tld
tokenizers
tomli
tomlkit
torch
tornado
tqdm
trafilatura
traitlets
trio
trio-websocket
trove-classifiers
typer
types-python-dateutil
typing-inspect
typing_extensions
tzdata
tzlocal
ujson
uri-template
uritemplate
urllib3
uv
uvicorn
virtualenv
watchfiles
wcwidth
webcolors
webencodings
websocket-client
websockets
widgetsnbextension
wrapt
wsproto
xxhash
y-py
yarl
ypy-websocket
zipp
